fix overflow

+++ Sources/Sentry/SentryDispatchFactory.m
@@ -19,8 +19,8 @@
     return [[SentryDispatchQueueWrapper alloc] initWithName:name attributes:attributes];
 }
 
-- (SentryDispatchSourceWrapper *)sourceWithInterval:(uint64_t)interval
-                                             leeway:(uint64_t)leeway
+- (SentryDispatchSourceWrapper *)sourceWithInterval:(NSInteger)interval
+                                             leeway:(NSInteger)leeway
                                           queueName:(const char *)queueName
                                          attributes:(dispatch_queue_attr_t)attributes
                                        eventHandler:(void (^)(void))eventHandler
